rename interface and result (was only partially renamed earlier)
