+++ src/Core/Domain/Product/SpecificPrice/QueryHandler/GetSpecificPriceListHandlerInterface.php
@@ -26,48 +26,17 @@
 
 declare(strict_types=1);
 
-namespace PrestaShop\PrestaShop\Core\Domain\Product\SpecificPrice\QueryResult;
+namespace PrestaShop\PrestaShop\Core\Domain\Product\SpecificPrice\QueryHandler;
 
-/**
- * Transfer SpecificPrice list data
- */
-class SpecificPriceListForEditing
+use PrestaShop\PrestaShop\Core\Domain\Product\SpecificPrice\Query\GetSpecificPriceList;
+use PrestaShop\PrestaShop\Core\Domain\Product\SpecificPrice\QueryResult\SpecificPriceList;
+
+interface GetSpecificPriceListHandlerInterface
 {
     /**
-     * @var SpecificPriceForListing[]
+     * @param GetSpecificPriceList $query
+     *
+     * @return SpecificPriceList
      */
-    private $specificPrices;
-
-    /**
-     * @var int
-     */
-    private $totalSpecificPricesCount;
-
-    /**
-     * @param SpecificPriceForListing[] $specificPrices
-     * @param int $totalSpecificPricesCount
-     */
-    public function __construct(
-        array $specificPrices,
-        int $totalSpecificPricesCount
-    ) {
-        $this->specificPrices = $specificPrices;
-        $this->totalSpecificPricesCount = $totalSpecificPricesCount;
-    }
-
-    /**
-     * @return SpecificPriceForListing[]
-     */
-    public function getSpecificPrices(): array
-    {
-        return $this->specificPrices;
-    }
-
-    /**
-     * @return int
-     */
-    public function getTotalSpecificPricesCount(): int
-    {
-        return $this->totalSpecificPricesCount;
-    }
+    public function handle(GetSpecificPriceList $query): SpecificPriceList;
 }
